\name{forgelib-metadata}
\alias{forgelib_status}
\alias{forgelib_repository}
\alias{forgelib_notice}
\title{ForgeLib package metadata}
\description{
Metadata helpers for the placeholder ForgeLib R package.
}
\usage{
forgelib_status()
forgelib_repository()
forgelib_notice()
}
\value{
A length-one character vector.
}
\details{
This release reserves the first-party R package name for the future Forge binding.
}
