% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molds.R
\name{mold}
\alias{mold}
\alias{mold_integer}
\alias{mold_scalar_integer}
\alias{mold_double}
\alias{mold_scalar_double}
\alias{mold_character}
\alias{mold_scalar_character}
\alias{mold_choice}
\title{Mold values into shape}
\usage{
mold_integer(x, n = NULL, allow_na = FALSE, allow_null = FALSE)

mold_scalar_integer(x, allow_na = FALSE, allow_null = FALSE)

mold_double(x, n = NULL, allow_na = FALSE, allow_null = FALSE)

mold_scalar_double(x, allow_na = FALSE, allow_null = FALSE)

mold_character(x, n = NULL, allow_na = FALSE, allow_null = FALSE)

mold_scalar_character(x, allow_na = FALSE, allow_null = FALSE)

mold_choice(x, choices, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{n}{The required length of the vector. If \code{NULL}, the length is not checked.}

\item{allow_na}{Whether to allow \code{NA}s in the vector.}

\item{allow_null}{Whether to allow \code{NULL}.}

\item{choices}{A character, numeric, or integer vector of allowed values.}
}
\description{
These functions verify and attempt to coerce values into the specified types and shapes. If they
are unsuccessful in the coercion, an error is thrown.
}
