% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certify.R
\name{certify}
\alias{certify}
\alias{gt}
\alias{gte}
\alias{lt}
\alias{lte}
\alias{bounded}
\title{Ensure Conditions on a Value}
\usage{
certify(.x, ..., .id = NULL)

gt(.l)

gte(.l)

lt(.u)

lte(.u)

bounded(.l = NULL, .u = NULL, .incl_lower = FALSE,
  .incl_upper = FALSE)
}
\arguments{
\item{.x}{The value to be checked.}

\item{...}{Conditions to be checked; should be functions that return TRUE/FALSE.}

\item{.id}{Name given to the input to aid the user in identifying the bad value.}

\item{.l}{Lower bound for the inequality condition.}

\item{.u}{Upper bound for the inequality condition.}

\item{.incl_lower}{Whether to include the left endpoint.}

\item{.incl_upper}{Whether to include the right endpoint.}
}
\description{
Checks that the input value satisfies specified conditions
}
